+++ src/handler/http/router/openapi.rs
@@ -149,9 +149,7 @@ use crate::{common::meta, handler::http::request};
             meta::dashboards::v1::QueryData,
             meta::dashboards::v1::CustomFieldsOption,
             meta::dashboards::v1::VariableList,
-            meta::dashboards::Folder,
             meta::dashboards::MoveDashboard,
-            meta::dashboards::FolderList,
             config::meta::alerts::alert::Alert,
             config::meta::alerts::Aggregation,
             config::meta::alerts::AggFunction,
@@ -167,6 +165,8 @@ use crate::{common::meta, handler::http::request};
             config::meta::alerts::QueryCondition,
             config::meta::alerts::TriggerCondition,
             config::meta::alerts::templates::Template,
+            config::meta::folder::Folder,
+            config::meta::folder::FolderList,
             config::meta::function::Transform,
             config::meta::function::FunctionList,
             config::meta::function::StreamOrder,
